drm/nouveau/ltc: fix tag base address getting truncated if above 4GiB
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
Rebuild_CHGLOG: - [drm] nouveau/ltc: fix tag base address getting truncated if above 4GiB (Rob Clark) [1173317]
Rebuild_FUZZ: 97.01%
commit-author Ben Skeggs <bskeggs@redhat.com>
commit 147ed897e807e1a6e2c6fd9dc91830f1d2831d43

(cherry picked from commit 147ed897e807e1a6e2c6fd9dc91830f1d2831d43)

+++ drivers/gpu/drm/nouveau/core/subdev/ltcg/nvc0.c
@@ -144,7 +144,7 @@ nvc0_ltcg_init_tag_ram(struct nouveau_fb *pfb, struct nvc0_ltcg_priv *priv)
 	if (ret) {
 		priv->num_tags = 0;
 	} else {
-		u64 tag_base = (priv->tag_ram->offset << 12) + tag_margin;
+		u64 tag_base = ((u64)priv->tag_ram->offset << 12) + tag_margin;
 
 		tag_base += tag_align - 1;
 		ret = do_div(tag_base, tag_align);
